refactor(kftray-kube): split tcp_forwarder by relay mode
Split kube/tcp_forwarder.rs (977 LOC) into kube/tcp_forwarder/:
- mod.rs: TcpForwarder struct, socket opts, forward_connection,
  forward_tls_streams, simple relay + tests
- logging.rs: HTTP-aware relay (forward_client_to_upstream,
  forward_upstream_to_client), ResponseState, response analysis + tests

Tests travel with their code. Zero behaviour change.
